refactor: Aktuelle Aufsicht page.tsx entflechten (#2421)
- useSupervisionDashboard-Hook: aggregierter Fetch (403-Einmal-Retry),
  Selektions-Reconciliation und Berlin-Mitternachts-Rollover (Kalendertag
  als Key-Suffix) gekapselt; die Page konsumiert nur noch das Ergebnis.
- SWR->useState-Kopierschicht eliminiert: students, allRooms,
  schulhofStatus, plannedNow, Gruppen-Maps, Capabilities usw. sind jetzt
  useMemo-Ableitungen eines einzigen dokumentierten Snapshots (Render-
  Adjust-Pattern statt Effekt+setState pro Feld).
- Ref-Loop-Guards abgebaut: lastDashboardReconciliationRef und
  previousDashboardDayRef entfallen; requestedGroupIdRef bleibt als
  Request-Parameter, weil der SSE-Vertrag den stabilen Key
  active-supervision-dashboard-* erzwingt (mutate-basierte Steuerung).
  Der Reconciliation-Effekt feuert nur noch auf Wertaenderungen von
  gewuenschter/aufgeloester Session und kann nicht mehr schleifen.
- Page in Sektionskomponenten und Hooks gesplittet (Roster, Kinder-Grid,
  Header, Reopen-Banner, Filter, Timetable-/Schulhof-Aktionen);
  page.tsx: 2841 -> 553 Zeilen. Tote Mirror-Refs (currentRoomRef,
  hasSupervisionRef, groupNameToIdMapRef) entfernt.
- SSE-Key-Praefix, 403-Retry ohne group_id und Roster-Keys unveraendert;
  alle 328 Seitentests und die use-global-sse-Tests laufen unveraendert
  gruen.
