+++ frontend/src/components/active-supervisions/timetable-roster.tsx
@@ -1,0 +1,667 @@
+"use client";
+
+import { useState } from "react";
+import { UserPlus } from "lucide-react";
+import { MotoConceptIcon } from "~/components/ui/moto-concept-icon";
+import { Button } from "~/components/ui/button";
+import {
+  clearOwnAttendanceMutation,
+  markOwnAttendanceMutation,
+} from "~/lib/sse-optimistic-mutations";
+import { useMinuteClock } from "~/lib/pickup-helpers";
+import { canCompleteInstance } from "~/lib/timetable-lifecycle";
+import { timetableOperationsApi } from "~/lib/timetable-operations-api";
+import type {
+  TimetableRoster,
+  TimetableRosterRow,
+} from "~/lib/timetable-operations-types";
+import { isCareDayExpected, isNotScheduledRow } from "~/lib/timetable-types";
+import type { Student } from "~/lib/student-helpers";
+
+/**
+ * Marks the attendance mutation as our own before firing the request so the
+ * global SSE handler can skip the echo event, and clears the mark again when
+ * the request fails (the echo will never arrive).
+ */
+export async function runOwnAttendanceMutation<T>(
+  eventType: "student_checkin" | "student_checkout",
+  studentId: string,
+  request: () => Promise<T>,
+): Promise<T> {
+  markOwnAttendanceMutation(eventType, studentId);
+  try {
+    return await request();
+  } catch (error) {
+    clearOwnAttendanceMutation(eventType, studentId);
+    throw error;
+  }
+}
+
+const ATTENDANCE_SUBSTATUS_LABELS: Record<
+  NonNullable<TimetableRosterRow["substatus"]>,
+  string
+> = {
+  late: "Verspätet",
+  excused: "Entschuldigt",
+  sick: "Krank",
+  field_trip: "Ausflug",
+  other: "Sonstiges",
+};
+
+// Builds the info notice after a check-in auto-moved the child out of another
+// running session (#2386). Null when the response reports no move.
+export function moveNoticeFromRoster(
+  roster: TimetableRoster,
+  studentId: string,
+): string | null {
+  if (roster.movedFrom == null) return null;
+  const name =
+    roster.rows.find((row) => row.studentId === studentId)?.studentName ??
+    "Das Kind";
+  return roster.movedFrom
+    ? `${name} wurde aus „${roster.movedFrom}“ hierher geholt.`
+    : `${name} wurde aus einer anderen Aktivität hierher geholt.`;
+}
+
+function rosterStudentMeta(
+  row: TimetableRosterRow,
+  instanceIsSpontaneous: boolean,
+): string {
+  return [
+    row.schoolClass,
+    row.groupName,
+    row.isUnplanned && !instanceIsSpontaneous ? "ungeplant" : "",
+  ]
+    .filter((value, index, values) => value && values.indexOf(value) === index)
+    .join(" · ");
+}
+
+function RosterSummaryStat({
+  label,
+  value,
+}: Readonly<{ label: string; value: number }>) {
+  return (
+    <div className="rounded-xl bg-white/80 px-3 py-2 shadow-[0_1px_0_rgba(17,24,39,0.04)]">
+      <span className="block text-sm font-semibold text-gray-900">{value}</span>
+      <span className="block text-[11px] font-medium text-gray-500">
+        {label}
+      </span>
+    </div>
+  );
+}
+
+export type RosterAction =
+  "check-in" | "check-out" | "excused" | "absent" | "expected";
+
+export async function runRosterActionRequest(
+  action: RosterAction,
+  instanceId: string,
+  studentId: string,
+): Promise<TimetableRoster | null> {
+  if (action === "check-in") {
+    return runOwnAttendanceMutation("student_checkin", studentId, () =>
+      timetableOperationsApi.checkIn(instanceId, studentId),
+    );
+  }
+  if (action === "check-out") {
+    return runOwnAttendanceMutation("student_checkout", studentId, () =>
+      timetableOperationsApi.checkOut(instanceId, studentId),
+    );
+  }
+  if (action === "expected") {
+    await timetableOperationsApi.patchAttendance(instanceId, studentId, {
+      status: "expected",
+      substatus: null,
+      note: null,
+    });
+    return null;
+  }
+  await timetableOperationsApi.patchAttendance(
+    instanceId,
+    studentId,
+    action === "excused"
+      ? { status: "absent", substatus: "excused" }
+      : { status: "absent" },
+  );
+  return null;
+}
+
+interface RosterRowActionsProps {
+  readonly row: TimetableRosterRow;
+  readonly onAction: (
+    action: RosterAction,
+    row: TimetableRosterRow,
+  ) => Promise<void>;
+}
+
+function RosterRowActions({ row, onAction }: RosterRowActionsProps) {
+  const runAction = async (action: RosterAction) => {
+    await onAction(action, row);
+  };
+
+  return (
+    <div className="flex flex-wrap gap-2">
+      {!row.currentlyPresent && row.status === "expected" ? (
+        <button
+          type="button"
+          onClick={() => runAction("check-in")}
+          className="bg-moto-green rounded-md px-3 py-2 text-sm font-medium text-gray-950"
+        >
+          Einchecken
+        </button>
+      ) : null}
+      {!row.currentlyPresent && row.status !== "expected" ? (
+        <button
+          type="button"
+          onClick={() => runAction("check-in")}
+          className="bg-moto-green rounded-md px-3 py-2 text-sm font-medium text-gray-950"
+        >
+          Wieder einchecken
+        </button>
+      ) : null}
+      {row.currentlyPresent ? (
+        <button
+          type="button"
+          onClick={() => runAction("check-out")}
+          className="rounded-md border border-gray-300 px-3 py-2 text-sm font-medium text-gray-700"
+        >
+          Raum verlassen
+        </button>
+      ) : null}
+      {row.planned &&
+      row.status === "expected" &&
+      isCareDayExpected(row.careDayStatus) ? (
+        <>
+          <button
+            type="button"
+            onClick={() => runAction("excused")}
+            className="border-moto-purple text-moto-purple-strong rounded-md border px-3 py-2 text-sm font-medium"
+          >
+            Entschuldigt
+          </button>
+          <button
+            type="button"
+            onClick={() => runAction("absent")}
+            className="rounded-md border border-gray-300 px-3 py-2 text-sm font-medium text-gray-700"
+          >
+            Abwesend
+          </button>
+        </>
+      ) : null}
+      {row.planned && !row.currentlyPresent && row.status === "absent" ? (
+        <button
+          type="button"
+          onClick={() => runAction("expected")}
+          className="rounded-md border border-gray-300 px-3 py-2 text-sm font-medium text-gray-700"
+        >
+          Zurück auf erwartet
+        </button>
+      ) : null}
+    </div>
+  );
+}
+
+interface TimetableRosterRowProps {
+  readonly attendanceWebEnabled: boolean;
+  readonly instanceIsSpontaneous: boolean;
+  readonly row: TimetableRosterRow;
+  readonly onAction: RosterRowActionsProps["onAction"];
+}
+
+function TimetableRosterStudentRow({
+  attendanceWebEnabled,
+  instanceIsSpontaneous,
+  row,
+  onAction,
+}: TimetableRosterRowProps) {
+  const attendanceDetail = [
+    row.substatus ? ATTENDANCE_SUBSTATUS_LABELS[row.substatus] : null,
+    row.note,
+  ]
+    .filter(Boolean)
+    .join(" · ");
+
+  return (
+    <div className="flex flex-col gap-3 border-b border-gray-100 px-4 py-3 last:border-b-0 sm:flex-row sm:items-center sm:justify-between">
+      <div>
+        <div className="font-medium text-gray-900">
+          {row.studentName || `Kind ${row.studentId}`}
+        </div>
+        <div className="mt-1 text-sm text-gray-500">
+          {rosterStudentMeta(row, instanceIsSpontaneous)}
+        </div>
+        {attendanceDetail ? (
+          <div className="text-moto-amber-strong mt-1 text-sm">
+            {attendanceDetail}
+          </div>
+        ) : null}
+        {row.parallelPresentIn ? (
+          <div className="text-moto-amber-strong mt-1 text-sm">
+            {`Auch in „${row.parallelPresentIn.title}“ (${row.parallelPresentIn.startTime}–${row.parallelPresentIn.endTime}) als anwesend eingetragen`}
+          </div>
+        ) : null}
+      </div>
+      {attendanceWebEnabled ? (
+        <RosterRowActions row={row} onAction={onAction} />
+      ) : null}
+    </div>
+  );
+}
+
+interface TimetableRosterSectionProps {
+  readonly attendanceWebEnabled: boolean;
+  readonly instanceIsSpontaneous: boolean;
+  readonly onAction: RosterRowActionsProps["onAction"];
+  readonly rows: TimetableRosterRow[];
+  readonly showTimetableCounts: boolean;
+  readonly title: string;
+}
+
+function TimetableRosterSection({
+  attendanceWebEnabled,
+  instanceIsSpontaneous,
+  onAction,
+  rows,
+  showTimetableCounts,
+  title,
+}: TimetableRosterSectionProps) {
+  if (rows.length === 0) return null;
+  const countLabel = showTimetableCounts ? ` (${rows.length})` : "";
+
+  return (
+    <section className="moto-content-surface overflow-hidden rounded-lg border">
+      <div className="border-b border-gray-100 bg-gray-50 px-4 py-2 text-sm font-semibold text-gray-700">
+        {title}
+        {countLabel}
+      </div>
+      {rows.map((row) => (
+        <TimetableRosterStudentRow
+          key={`${row.studentId}-${row.status}-${row.visitId ?? "planned"}`}
+          attendanceWebEnabled={attendanceWebEnabled}
+          instanceIsSpontaneous={instanceIsSpontaneous}
+          row={row}
+          onAction={onAction}
+        />
+      ))}
+    </section>
+  );
+}
+
+function confirmExpectedLabel(
+  isConfirmingExpected: boolean,
+  showTimetableCounts: boolean,
+  count: number,
+): string {
+  if (isConfirmingExpected) return "Bestätigt...";
+  if (showTimetableCounts) return `${count} erwartete bestätigen`;
+  return "Erwartete bestätigen";
+}
+
+interface TimetableRosterHeaderProps {
+  readonly attendanceWebEnabled: boolean;
+  readonly confirmableExpectedRows: TimetableRosterRow[];
+  readonly isCompletingInstance: boolean;
+  readonly isConfirmingExpected: boolean;
+  readonly roster: TimetableRoster;
+  readonly showTimetableCounts: boolean;
+  readonly summary: {
+    readonly absent: number;
+    readonly departed: number;
+    readonly expected: number;
+    readonly present: number;
+    readonly unplanned: number;
+  };
+  readonly onComplete: () => Promise<void>;
+  readonly onConfirmExpected: (rows: TimetableRosterRow[]) => Promise<void>;
+}
+
+function TimetableRosterHeader({
+  attendanceWebEnabled,
+  confirmableExpectedRows,
+  isCompletingInstance,
+  isConfirmingExpected,
+  roster,
+  showTimetableCounts,
+  summary,
+  onComplete,
+  onConfirmExpected,
+}: TimetableRosterHeaderProps) {
+  const now = useMinuteClock();
+  const completeEnabled = canCompleteInstance(
+    roster.instance.canComplete,
+    roster.instance.completeAvailableAt,
+    now,
+  );
+  const handleConfirmExpectedClick = async () => {
+    await onConfirmExpected(confirmableExpectedRows);
+  };
+  const handleCompleteClick = async () => {
+    await onComplete();
+  };
+  const confirmLabel = confirmExpectedLabel(
+    isConfirmingExpected,
+    showTimetableCounts,
+    confirmableExpectedRows.length,
+  );
+
+  return (
+    <div className="moto-content-surface overflow-hidden rounded-2xl border border-gray-200 shadow-sm backdrop-blur-md">
+      <div className="flex flex-col gap-3 border-b border-gray-100 p-4 sm:flex-row sm:items-center sm:justify-between">
+        <div className="flex min-w-0 items-center gap-3">
+          <span className="flex h-9 w-9 shrink-0 items-center justify-center rounded-xl bg-gray-100">
+            <MotoConceptIcon concept="present" size={18} />
+          </span>
+          <div className="min-w-0">
+            <p className="text-moto-green-strong text-xs font-semibold tracking-wide uppercase">
+              Aktiv
+            </p>
+            <h2 className="truncate text-base font-semibold text-gray-900">
+              {roster.instance.title}
+            </h2>
+            <p className="truncate text-sm text-gray-600">
+              {roster.instance.roomName ?? `Raum ${roster.instance.roomId}`} ·{" "}
+              {roster.instance.startTime}-{roster.instance.endTime}
+            </p>
+          </div>
+        </div>
+        <div className="flex flex-wrap gap-2 sm:justify-end">
+          {attendanceWebEnabled ? (
+            <button
+              type="button"
+              disabled={
+                isConfirmingExpected || confirmableExpectedRows.length === 0
+              }
+              onClick={handleConfirmExpectedClick}
+              className="bg-moto-green hover:bg-moto-green-hover focus-visible:ring-moto-green/30 inline-flex h-9 items-center justify-center gap-2 rounded-lg px-3 text-sm font-medium text-gray-950 shadow-sm transition-colors focus-visible:ring-2 focus-visible:outline-none disabled:cursor-not-allowed disabled:opacity-50"
+            >
+              <MotoConceptIcon concept="present" size={16} />
+              {confirmLabel}
+            </button>
+          ) : null}
+          {attendanceWebEnabled ? (
+            <button
+              type="button"
+              disabled={isCompletingInstance || !completeEnabled}
+              onClick={handleCompleteClick}
+              className="inline-flex h-9 items-center justify-center rounded-lg bg-gray-900 px-3 text-sm font-medium text-white shadow-sm transition-colors hover:bg-gray-700 focus-visible:ring-2 focus-visible:ring-gray-400 focus-visible:outline-none disabled:opacity-50"
+            >
+              {completeEnabled
+                ? "Beenden"
+                : `Beenden ab ${roster.instance.endTime}`}
+            </button>
+          ) : null}
+        </div>
+      </div>
+      {showTimetableCounts ? (
+        <div className="grid grid-cols-2 gap-2 p-4 sm:grid-cols-5">
+          <RosterSummaryStat label="Anwesend" value={summary.present} />
+          <RosterSummaryStat label="Erwartet" value={summary.expected} />
+          <RosterSummaryStat label="Abwesend" value={summary.absent} />
+          <RosterSummaryStat label="Gegangen" value={summary.departed} />
+          <RosterSummaryStat label="Ungeplant" value={summary.unplanned} />
+        </div>
+      ) : null}
+    </div>
+  );
+}
+
+interface AddUnplannedStudentFormProps {
+  readonly isAddingStudent: boolean;
+  readonly results: Student[];
+  readonly search: string;
+  readonly onAdd: (studentId: string) => Promise<boolean>;
+  readonly onSearchChange: (value: string) => void;
+}
+
+function AddUnplannedStudentForm({
+  isAddingStudent,
+  results,
+  search,
+  onAdd,
+  onSearchChange,
+}: AddUnplannedStudentFormProps) {
+  const [selectedId, setSelectedId] = useState<string | null>(null);
+  // Derived against the current results so a stale selection from a previous
+  // search can never add the wrong child.
+  const selectedStudent =
+    results.find((student) => student.id.toString() === selectedId) ?? null;
+  const targetStudent =
+    selectedStudent ?? (results.length === 1 ? (results[0] ?? null) : null);
+  const addStudent = async (studentId: string) => {
+    if (await onAdd(studentId)) {
+      setSelectedId(null);
+    }
+  };
+  const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    if (targetStudent && !isAddingStudent) {
+      await addStudent(targetStudent.id.toString());
+    }
+  };
+
+  return (
+    <form
+      className="moto-content-surface rounded-2xl border p-4 shadow-sm"
+      onSubmit={handleSubmit}
+    >
+      <div className="mb-3 flex items-center gap-2 text-sm font-semibold text-gray-900">
+        <UserPlus
+          className="text-moto-green-vivid h-4 w-4"
+          aria-hidden="true"
+        />
+        Kind ungeplant hinzufügen
+      </div>
+      <div className="flex flex-col gap-2 sm:flex-row">
+        <input
+          type="search"
+          name="unplanned-student-search"
+          aria-label="Kind ungeplant suchen"
+          value={search}
+          onChange={(event) => {
+            setSelectedId(null);
+            onSearchChange(event.target.value);
+          }}
+          placeholder="Weiteres Kind suchen..."
+          className="focus:border-moto-green focus:ring-moto-green/20 min-h-10 flex-1 rounded-lg border border-gray-300 px-3 text-sm focus:ring-2 focus:outline-none"
+        />
+        <button
+          type="submit"
+          disabled={isAddingStudent || !targetStudent}
+          className="bg-moto-green hover:bg-moto-green-hover rounded-lg px-4 py-2 text-sm font-medium text-gray-950 shadow-sm transition-colors disabled:opacity-50"
+        >
+          Hinzufügen
+        </button>
+      </div>
+      {results.length > 0 ? (
+        <div className="mt-2 grid gap-2 sm:grid-cols-2">
+          {results.map((student) => {
+            const studentId = student.id.toString();
+            const isSelected = selectedStudent?.id.toString() === studentId;
+            return (
+              <Button
+                key={student.id}
+                type="button"
+                variant="ghost"
+                size="md"
+                disabled={isAddingStudent}
+                aria-pressed={isSelected}
+                onClick={() =>
+                  setSelectedId((prev) =>
+                    prev === studentId ? null : studentId,
+                  )
+                }
+                className={`min-h-11 w-full !justify-start border px-3 text-left !shadow-none ${
+                  isSelected
+                    ? "!border-moto-green !bg-moto-green/10 hover:!bg-moto-green/15"
+                    : "hover:!border-moto-green !border-gray-200 !bg-transparent hover:!bg-gray-100"
+                }`}
+              >
+                <span className="font-medium text-gray-900">
+                  {student.name ||
+                    [student.first_name, student.second_name]
+                      .filter(Boolean)
+                      .join(" ")}
+                </span>
+                <span className="ml-2 text-gray-500">
+                  {[student.school_class, student.group_name]
+                    .filter(Boolean)
+                    .join(" · ")}
+                </span>
+              </Button>
+            );
+          })}
+        </div>
+      ) : null}
+      {results.length > 1 && !selectedStudent ? (
+        <p className="mt-2 text-sm text-gray-500">
+          Bitte ein Kind aus der Liste antippen.
+        </p>
+      ) : null}
+    </form>
+  );
+}
+
+interface TimetableRosterContentProps {
+  readonly addStudentResults: Student[];
+  readonly addStudentSearch: string;
+  readonly attendanceWebEnabled: boolean;
+  readonly isAddingStudent: boolean;
+  readonly isCompletingInstance: boolean;
+  readonly isConfirmingExpected: boolean;
+  readonly roster: TimetableRoster;
+  readonly showTimetableCounts: boolean;
+  readonly onAddStudent: (studentId: string) => Promise<boolean>;
+  readonly onComplete: () => Promise<void>;
+  readonly onConfirmExpected: (rows: TimetableRosterRow[]) => Promise<void>;
+  readonly onRosterAction: RosterRowActionsProps["onAction"];
+  readonly onSearchChange: (value: string) => void;
+}
+
+export function TimetableRosterContent({
+  addStudentResults,
+  addStudentSearch,
+  attendanceWebEnabled,
+  isAddingStudent,
+  isCompletingInstance,
+  isConfirmingExpected,
+  roster,
+  showTimetableCounts,
+  onAddStudent,
+  onComplete,
+  onConfirmExpected,
+  onRosterAction,
+  onSearchChange,
+}: TimetableRosterContentProps) {
+  const present = roster.rows.filter(
+    (row) => row.currentlyPresent && row.planned,
+  );
+  // The care plan decides who counts as expected (#1747): rows the plan does
+  // not place here today — not booked, or the day was cancelled — go into
+  // their own section below, never into "Erwartet" and never into the bulk
+  // confirm, which would persist attendance for a child who is not coming.
+  const expected = roster.rows.filter(
+    (row) =>
+      row.planned &&
+      !row.currentlyPresent &&
+      row.status === "expected" &&
+      isCareDayExpected(row.careDayStatus),
+  );
+  // An absence a sick / excused / class-trip day status wrote onto a day the
+  // child was never booked into care belongs here too, not under "Abwesend":
+  // the block has not ended yet, so nothing has undone that false absence, and
+  // the header count already groups it this way (#1747).
+  const notScheduled = roster.rows.filter(
+    (row) =>
+      row.planned &&
+      !row.currentlyPresent &&
+      isNotScheduledRow(row.status, row.careDayStatus),
+  );
+  const absent = roster.rows.filter(
+    (row) =>
+      row.planned &&
+      !row.currentlyPresent &&
+      row.status === "absent" &&
+      !isNotScheduledRow(row.status, row.careDayStatus),
+  );
+  const departed = roster.rows.filter(
+    (row) =>
+      !row.currentlyPresent &&
+      (row.status === "present" || (row.isUnplanned && row.visitId)),
+  );
+  const unplanned = roster.rows.filter(
+    (row) => row.isUnplanned && row.currentlyPresent,
+  );
+  const confirmableExpectedRows = expected.filter(
+    (row) => row.planned && !row.currentlyPresent,
+  );
+  const instanceIsSpontaneous = roster.instance.isSpontaneous;
+  const unplannedTitle = instanceIsSpontaneous ? "Teilnehmende" : "Ungeplant";
+  const sectionProps = {
+    attendanceWebEnabled,
+    instanceIsSpontaneous,
+    onAction: onRosterAction,
+    showTimetableCounts,
+  };
+
+  return (
+    <div className="space-y-4">
+      <TimetableRosterHeader
+        attendanceWebEnabled={attendanceWebEnabled}
+        confirmableExpectedRows={confirmableExpectedRows}
+        isCompletingInstance={isCompletingInstance}
+        isConfirmingExpected={isConfirmingExpected}
+        roster={roster}
+        showTimetableCounts={showTimetableCounts}
+        summary={{
+          absent: absent.length,
+          departed: departed.length,
+          expected: expected.length,
+          present: present.length,
+          unplanned: unplanned.length,
+        }}
+        onComplete={onComplete}
+        onConfirmExpected={onConfirmExpected}
+      />
+      {attendanceWebEnabled ? (
+        <AddUnplannedStudentForm
+          key={roster.instance.id}
+          isAddingStudent={isAddingStudent}
+          results={addStudentResults}
+          search={addStudentSearch}
+          onAdd={onAddStudent}
+          onSearchChange={onSearchChange}
+        />
+      ) : null}
+      <TimetableRosterSection
+        title="Anwesend"
+        rows={present}
+        {...sectionProps}
+      />
+      <TimetableRosterSection
+        title="Erwartet"
+        rows={expected}
+        {...sectionProps}
+      />
+      <TimetableRosterSection
+        title="Heute nicht eingeplant"
+        rows={notScheduled}
+        {...sectionProps}
+      />
+      <TimetableRosterSection
+        title="Entschuldigt / Abwesend"
+        rows={absent}
+        {...sectionProps}
+      />
+      <TimetableRosterSection
+        title="Nicht mehr im Raum"
+        rows={departed}
+        {...sectionProps}
+      />
+      <TimetableRosterSection
+        title={unplannedTitle}
+        rows={unplanned}
+        {...sectionProps}
+      />
+    </div>
+  );
+}
